analyzer/requirements.txt
textblob
textblob_de
kafka-python
pyopenssl
pytest
pymongo
langdetect